The problem of god files is eliminated.
Average cyclomatic complexity by class down .05
