+++ Source/Client/AbstractRequest.php
@@ -1,0 +1,149 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PagoFacil\Payment\Source\Client;
+
+use Psr\Http\Message\RequestInterface;
+use Psr\Http\Message\StreamInterface;
+use Psr\Http\Message\UriInterface;
+
+class AbstractRequest implements RequestInterface
+{
+
+    /**
+     * @inheritDoc
+     */
+    public function getProtocolVersion()
+    {
+        // TODO: Implement getProtocolVersion() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function withProtocolVersion($version)
+    {
+        // TODO: Implement withProtocolVersion() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getHeaders()
+    {
+        // TODO: Implement getHeaders() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function hasHeader($name)
+    {
+        // TODO: Implement hasHeader() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getHeader($name)
+    {
+        // TODO: Implement getHeader() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getHeaderLine($name)
+    {
+        // TODO: Implement getHeaderLine() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function withHeader($name, $value)
+    {
+        // TODO: Implement withHeader() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function withAddedHeader($name, $value)
+    {
+        // TODO: Implement withAddedHeader() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function withoutHeader($name)
+    {
+        // TODO: Implement withoutHeader() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getBody()
+    {
+        // TODO: Implement getBody() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function withBody(StreamInterface $body)
+    {
+        // TODO: Implement withBody() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getRequestTarget()
+    {
+        // TODO: Implement getRequestTarget() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function withRequestTarget($requestTarget)
+    {
+        // TODO: Implement withRequestTarget() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getMethod()
+    {
+        // TODO: Implement getMethod() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function withMethod($method)
+    {
+        // TODO: Implement withMethod() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getUri()
+    {
+        // TODO: Implement getUri() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function withUri(UriInterface $uri, $preserveHost = false)
+    {
+        // TODO: Implement withUri() method.
+    }
+}
