+++ Source/Client/AbstractResponse.php
@@ -1,0 +1,149 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PagoFacil\Payment\Source\Client;
+
+use PagoFacil\Payment\Source\Client\Interfaces\PagoFacilResponseInterface;
+use PagoFacil\Payment\Source\Transaction\Charge;
+use Psr\Http\Message\StreamInterface;
+
+class AbstractResponse implements PagoFacilResponseInterface
+{
+
+    /**
+     * @inheritDoc
+     */
+    public function getProtocolVersion()
+    {
+        // TODO: Implement getProtocolVersion() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function withProtocolVersion($version)
+    {
+        // TODO: Implement withProtocolVersion() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getHeaders()
+    {
+        // TODO: Implement getHeaders() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function hasHeader($name)
+    {
+        // TODO: Implement hasHeader() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getHeader($name)
+    {
+        // TODO: Implement getHeader() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getHeaderLine($name)
+    {
+        // TODO: Implement getHeaderLine() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function withHeader($name, $value)
+    {
+        // TODO: Implement withHeader() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function withAddedHeader($name, $value)
+    {
+        // TODO: Implement withAddedHeader() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function withoutHeader($name)
+    {
+        // TODO: Implement withoutHeader() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getBody()
+    {
+        // TODO: Implement getBody() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function withBody(StreamInterface $body)
+    {
+        // TODO: Implement withBody() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function validateAuthorized(): void
+    {
+        // TODO: Implement validateAuthorized() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getBodyToArray(): array
+    {
+        // TODO: Implement getBodyToArray() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getTransaction(): Charge
+    {
+        // TODO: Implement getTransaction() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getStatusCode()
+    {
+        // TODO: Implement getStatusCode() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function withStatus($code, $reasonPhrase = '')
+    {
+        // TODO: Implement withStatus() method.
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getReasonPhrase()
+    {
+        // TODO: Implement getReasonPhrase() method.
+    }
+}
