+++ Source/Interfaces/RegistryInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PagoFacil\Payment\Source\Interfaces;
+
+interface RegistryInterface
+{
+    public function get(string $key);
+    public function set(string $key, $value): void;
+    public function deleteKey(string $key): void;
+    public function deleteAll();
+}
